requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.2.0
autopep8==1.6.0
bidict==0.21.4
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
dnspython==2.1.0
eventlet==0.33.0
Flask==2.0.2
Flask-SocketIO==4.3.2
frozenlist==1.2.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-dotenv==0.20.0
python-engineio==3.14.2
python-socketio==4.6.1
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
Werkzeug==2.0.2
yarl==1.7.2
